example_app/requirements.txt
uvicorn
google-generativeai
python-dotenv
# pipeline-factory
